pyproject.toml
[tool.poetry]
name = "spy-cats"
version = "0.1.0"
description = ""
authors = ["dchnkoo <valera.dchn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.22"
fastapi = {extras = ["all"], version = "^0.115.5"}
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
